ReadableFileImpl uses int64_t

+++ include/AVEVA/RocksDB/Plugin/Azure/Impl/ReadableFileImpl.hpp
@@ -13,8 +13,8 @@ namespace AVEVA::RocksDB::Plugin::Azure::Impl
         std::string m_name;
         std::shared_ptr<Core::BlobClient> m_blobClient;
         std::shared_ptr<Core::FileCache> m_fileCache;
-        uint64_t m_offset;
-        uint64_t m_size;
+        int64_t m_offset;
+        int64_t m_size;
 
     public:
         ReadableFileImpl(std::string_view name,
@@ -22,13 +22,13 @@ namespace AVEVA::RocksDB::Plugin::Azure::Impl
             std::shared_ptr<Core::FileCache> fileCache);
 
         // NOTE: Increments m_offset
-        [[nodiscard]] uint64_t SequentialRead(size_t bytesToRead, char* buffer);
+        [[nodiscard]] int64_t SequentialRead(int64_t bytesToRead, char* buffer);
 
         // NOTE: Random so doesn't affect the sequential reads
-        [[nodiscard]] uint64_t RandomRead(uint64_t offset, size_t bytesToRead, char* buffer) const;
+        [[nodiscard]] int64_t RandomRead(int64_t offset, int64_t bytesToRead, char* buffer) const;
 
-        uint64_t GetOffset() const;
-        void Skip(uint64_t n);
-        uint64_t GetSize() const;
+        int64_t GetOffset() const;
+        void Skip(int64_t n);
+        int64_t GetSize() const;
     };
 }
