+++ tests/AVEVA/RocksDB/Plugin/Azure/Impl/ReadableFileTests.cpp
@@ -9,6 +9,7 @@ using AVEVA::RocksDB::Plugin::Azure::Impl::ReadableFileImpl;
 using AVEVA::RocksDB::Plugin::Azure::Impl::Configuration;
 using AVEVA::RocksDB::Plugin::Core::Mocks::BlobClientMock;
 using ::testing::_;
+
 using ::testing::Return;
 using ::testing::DoAll;
 using ::testing::SetArrayArgument;
@@ -16,7 +17,7 @@ using ::testing::SetArrayArgument;
 class ReadableFileTests : public ::testing::Test
 {
 protected:
-    std::shared_ptr<BlobClientMock> m_blobClient;
+  std::shared_ptr<BlobClientMock> m_blobClient;
     static const constexpr uint64_t DefaultBlobSize = Configuration::PageBlob::PageSize * 2;
 
     void SetUp() override
@@ -24,7 +25,7 @@ protected:
         m_blobClient = std::make_shared<BlobClientMock>();
 
         // Default behavior: return a blob size
-        ON_CALL(*m_blobClient, GetSize())
+      ON_CALL(*m_blobClient, GetSize())
             .WillByDefault(Return(DefaultBlobSize));
     }
 };
@@ -40,23 +41,23 @@ TEST_F(ReadableFileTests, Constructor_InitializesWithBlobSize)
     ReadableFileImpl file{ "test.sst", m_blobClient, nullptr };
 
     // Assert
-    EXPECT_EQ(expectedSize, file.GetSize());
+    EXPECT_EQ(static_cast<int64_t>(expectedSize), file.GetSize());
     EXPECT_EQ(0, file.GetOffset());
 }
 
 TEST_F(ReadableFileTests, SequentialRead_WithoutCache_ReadsFromBlob)
 {
     // Arrange
-    static const constexpr size_t bytesToRead = 100;
+    static const constexpr int64_t bytesToRead = 100;
     std::vector<char> buffer(bytesToRead);
     std::vector<char> expectedData(bytesToRead, 'A');
 
     EXPECT_CALL(*m_blobClient, DownloadTo(::testing::A<std::span<char>>(), 0, bytesToRead))
         .WillOnce([&expectedData](std::span<char> buffer, int64_t /*offset*/, int64_t /*length*/)
-            {
+       {
                 std::copy(expectedData.begin(), expectedData.end(), buffer.begin());
-                return static_cast<int64_t>(expectedData.size());
-            });
+              return static_cast<int64_t>(expectedData.size());
+   });
 
     ReadableFileImpl file{ "test.sst", m_blobClient, nullptr };
 
@@ -72,24 +73,24 @@ TEST_F(ReadableFileTests, SequentialRead_WithoutCache_ReadsFromBlob)
 TEST_F(ReadableFileTests, SequentialRead_MultipleReads_IncrementsOffset)
 {
     // Arrange
-    const constexpr size_t firstRead = 50;
-    const constexpr size_t secondRead = 75;
+    const constexpr int64_t firstRead = 50;
+    const constexpr int64_t secondRead = 75;
     std::vector<char> buffer1(firstRead);
     std::vector<char> buffer2(secondRead);
 
     EXPECT_CALL(*m_blobClient, DownloadTo(::testing::A<std::span<char>>(), 0, firstRead))
         .WillOnce([firstRead](std::span<char> buffer, int64_t /*offset*/, int64_t /*length*/)
             {
-                std::fill_n(buffer.begin(), firstRead, 'X');
-                return static_cast<int64_t>(firstRead);
-            });
+      std::fill_n(buffer.begin(), firstRead, 'X');
+       return static_cast<int64_t>(firstRead);
+       });
 
     EXPECT_CALL(*m_blobClient, DownloadTo(::testing::A<std::span<char>>(), firstRead, secondRead))
         .WillOnce([secondRead](std::span<char> buffer, int64_t /*offset*/, int64_t /*length*/)
             {
-                std::fill_n(buffer.begin(), secondRead, 'Y');
-                return static_cast<int64_t>(secondRead);
-            });
+  std::fill_n(buffer.begin(), secondRead, 'Y');
+     return static_cast<int64_t>(secondRead);
+          });
 
     ReadableFileImpl file{ "test.sst", m_blobClient, nullptr };
 
@@ -105,19 +106,19 @@ TEST_F(ReadableFileTests, SequentialRead_MultipleReads_IncrementsOffset)
 
 TEST_F(ReadableFileTests, SequentialRead_RequestMoreThanAvailable_ReadsOnlyAvailableBytes)
 {
-    // Arrange
+ // Arrange
     constexpr uint64_t blobSize = 100;
-    constexpr size_t bytesToRead = 150;
+    constexpr int64_t bytesToRead = 150;
     std::vector<char> buffer(bytesToRead);
 
     EXPECT_CALL(*m_blobClient, GetSize())
-        .WillOnce(Return(blobSize));
+   .WillOnce(Return(blobSize));
 
-    EXPECT_CALL(*m_blobClient, DownloadTo(::testing::A<std::span<char>>(), 0, blobSize))
+    EXPECT_CALL(*m_blobClient, DownloadTo(::testing::A<std::span<char>>(), 0, static_cast<int64_t>(blobSize)))
         .WillOnce([blobSize](std::span<char> /*buffer*/, int64_t /*offset*/, int64_t /*length*/)
             {
-                return static_cast<int64_t>(blobSize);
-            });
+           return static_cast<int64_t>(blobSize);
+       });
 
     ReadableFileImpl file{ "test.sst", m_blobClient, nullptr };
 
@@ -125,8 +126,8 @@ TEST_F(ReadableFileTests, SequentialRead_RequestMoreThanAvailable_ReadsOnlyAvail
     const auto bytesRead = file.SequentialRead(bytesToRead, buffer.data());
 
     // Assert
-    EXPECT_EQ(blobSize, bytesRead);
-    EXPECT_EQ(blobSize, file.GetOffset());
+    EXPECT_EQ(static_cast<int64_t>(blobSize), bytesRead);
+    EXPECT_EQ(static_cast<int64_t>(blobSize), file.GetOffset());
 }
 
 TEST_F(ReadableFileTests, SequentialRead_AtEndOfFile_ReturnsZero)
@@ -139,29 +140,29 @@ TEST_F(ReadableFileTests, SequentialRead_AtEndOfFile_ReturnsZero)
         .WillOnce(Return(blobSize));
 
     ReadableFileImpl file{ "test.sst", m_blobClient, nullptr };
-    file.Skip(blobSize); // Move to end of file
+file.Skip(static_cast<int64_t>(blobSize)); // Move to end of file
 
     // Act
     const auto bytesRead = file.SequentialRead(50, buffer.data());
 
     // Assert
-    EXPECT_EQ(0, bytesRead);
-    EXPECT_EQ(blobSize, file.GetOffset());
+  EXPECT_EQ(0, bytesRead);
+    EXPECT_EQ(static_cast<int64_t>(blobSize), file.GetOffset());
 }
 
 TEST_F(ReadableFileTests, RandomRead_WithoutCache_ReadsFromBlob)
 {
     // Arrange
-    constexpr uint64_t offset = 50;
-    constexpr size_t bytesToRead = 100;
+    constexpr int64_t offset = 50;
+    constexpr int64_t bytesToRead = 100;
     std::vector<char> buffer(bytesToRead);
     std::vector<char> expectedData(bytesToRead, 'B');
 
     EXPECT_CALL(*m_blobClient, DownloadTo(::testing::A<std::span<char>>(), offset, bytesToRead))
-        .WillOnce([&expectedData](std::span<char> buffer, int64_t /*offset*/, int64_t /*length*/)
+   .WillOnce([&expectedData](std::span<char> buffer, int64_t /*offset*/, int64_t /*length*/)
             {
                 std::copy(expectedData.begin(), expectedData.end(), buffer.begin());
-                return static_cast<int64_t>(expectedData.size());
+      return static_cast<int64_t>(expectedData.size());
             });
 
     ReadableFileImpl file{ "test.sst", m_blobClient, nullptr };
@@ -172,15 +173,15 @@ TEST_F(ReadableFileTests, RandomRead_WithoutCache_ReadsFromBlob)
     // Assert
     EXPECT_EQ(bytesToRead, bytesRead);
     EXPECT_EQ(expectedData, buffer);
-    EXPECT_EQ(0, file.GetOffset()); // Random read shouldn't affect sequential offset
+  EXPECT_EQ(0, file.GetOffset()); // Random read shouldn't affect sequential offset
 }
 
 TEST_F(ReadableFileTests, RandomRead_DoesNotAffectSequentialOffset)
 {
     // Arrange
-    constexpr size_t sequentialBytes = 50;
-    constexpr uint64_t randomOffset = 200;
-    constexpr size_t randomBytes = 75;
+    constexpr int64_t sequentialBytes = 50;
+    constexpr int64_t randomOffset = 200;
+    constexpr int64_t randomBytes = 75;
     std::vector<char> seqBuffer(sequentialBytes);
     std::vector<char> randomBuffer(randomBytes);
 
@@ -191,14 +192,14 @@ TEST_F(ReadableFileTests, RandomRead_DoesNotAffectSequentialOffset)
         .WillOnce(Return(static_cast<int64_t>(randomBytes)));
 
     EXPECT_CALL(*m_blobClient, DownloadTo(::testing::A<std::span<char>>(), sequentialBytes, sequentialBytes))
-        .WillOnce(Return(static_cast<int64_t>(sequentialBytes)));
+ .WillOnce(Return(static_cast<int64_t>(sequentialBytes)));
 
-    ReadableFileImpl file{ "test.sst", m_blobClient, nullptr };
+  ReadableFileImpl file{ "test.sst", m_blobClient, nullptr };
 
     // Act
     [[maybe_unused]] const auto bytesRead1 = file.SequentialRead(sequentialBytes, seqBuffer.data());
     [[maybe_unused]] const auto bytesRead2 = file.RandomRead(randomOffset, randomBytes, randomBuffer.data());
-    [[maybe_unused]] const auto bytesRead3 = file.SequentialRead(sequentialBytes, seqBuffer.data());
+[[maybe_unused]] const auto bytesRead3 = file.SequentialRead(sequentialBytes, seqBuffer.data());
 
     // Assert
     EXPECT_EQ(sequentialBytes * 2, file.GetOffset());
@@ -208,20 +209,20 @@ TEST_F(ReadableFileTests, RandomRead_RequestMoreThanAvailable_ReadsOnlyAvailable
 {
     // Arrange
     constexpr uint64_t blobSize = 200;
-    constexpr uint64_t offset = 150;
-    constexpr size_t bytesToRead = 100;
-    constexpr size_t expectedBytes = 50; // Only 50 bytes available from offset 150 in a 200 byte blob
+    constexpr int64_t offset = 150;
+    constexpr int64_t bytesToRead = 100;
+    constexpr int64_t expectedBytes = 50; // Only 50 bytes available from offset 150 in a 200 byte blob
     std::vector<char> buffer(bytesToRead);
 
     EXPECT_CALL(*m_blobClient, GetSize())
-        .WillOnce(Return(blobSize));
+  .WillOnce(Return(blobSize));
 
     EXPECT_CALL(*m_blobClient, DownloadTo(::testing::A<std::span<char>>(), offset, expectedBytes))
-        .WillOnce([expectedBytes](std::span<char> buffer, int64_t /*offset*/, int64_t /*length*/)
-            {
-                std::fill_n(buffer.begin(), expectedBytes, 'C');
-                return static_cast<int64_t>(expectedBytes);
-            });
+   .WillOnce([expectedBytes](std::span<char> buffer, int64_t /*offset*/, int64_t /*length*/)
+       {
+         std::fill_n(buffer.begin(), expectedBytes, 'C');
+           return static_cast<int64_t>(expectedBytes);
+});
 
     ReadableFileImpl file{ "test.sst", m_blobClient, nullptr };
 
@@ -244,7 +245,7 @@ TEST_F(ReadableFileTests, RandomRead_AtEndOfFile_ReturnsZero)
     ReadableFileImpl file{ "test.sst", m_blobClient, nullptr };
 
     // Act
-    const auto bytesRead = file.RandomRead(blobSize, 50, buffer.data());
+    const auto bytesRead = file.RandomRead(static_cast<int64_t>(blobSize), 50, buffer.data());
 
     // Assert
     EXPECT_EQ(0, bytesRead);
@@ -253,7 +254,7 @@ TEST_F(ReadableFileTests, RandomRead_AtEndOfFile_ReturnsZero)
 TEST_F(ReadableFileTests, Skip_IncrementsOffset)
 {
     // Arrange
-    constexpr uint64_t skipAmount = 100;
+    constexpr int64_t skipAmount = 100;
     ReadableFileImpl file{ "test.sst", m_blobClient, nullptr };
 
     // Act
@@ -266,14 +267,14 @@ TEST_F(ReadableFileTests, Skip_IncrementsOffset)
 TEST_F(ReadableFileTests, Skip_Multiple_AccumulatesOffset)
 {
     // Arrange
-    constexpr uint64_t skip1 = 50;
-    constexpr uint64_t skip2 = 75;
-    constexpr uint64_t skip3 = 25;
+  constexpr int64_t skip1 = 50;
+    constexpr int64_t skip2 = 75;
+    constexpr int64_t skip3 = 25;
     ReadableFileImpl file{ "test.sst", m_blobClient, nullptr };
 
     // Act
-    file.Skip(skip1);
-    file.Skip(skip2);
+  file.Skip(skip1);
+  file.Skip(skip2);
     file.Skip(skip3);
 
     // Assert
@@ -291,7 +292,7 @@ TEST_F(ReadableFileTests, GetOffset_InitiallyZero)
 
 TEST_F(ReadableFileTests, GetSize_ReturnsCorrectSize)
 {
-    // Arrange
+  // Arrange
     constexpr uint64_t expectedSize = 5000;
     EXPECT_CALL(*m_blobClient, GetSize())
         .WillOnce(Return(expectedSize));
@@ -300,13 +301,13 @@ TEST_F(ReadableFileTests, GetSize_ReturnsCorrectSize)
     ReadableFileImpl file{ "test.sst", m_blobClient, nullptr };
 
     // Assert
-    EXPECT_EQ(expectedSize, file.GetSize());
+    EXPECT_EQ(static_cast<int64_t>(expectedSize), file.GetSize());
 }
 
 TEST_F(ReadableFileTests, SequentialRead_DownloadReturnsNegative_ReturnsZero)
 {
     // Arrange
-    constexpr size_t bytesToRead = 100;
+    constexpr int64_t bytesToRead = 100;
     std::vector<char> buffer(bytesToRead);
 
     EXPECT_CALL(*m_blobClient, DownloadTo(::testing::A<std::span<char>>(), 0, bytesToRead))
@@ -325,30 +326,30 @@ TEST_F(ReadableFileTests, SequentialRead_DownloadReturnsNegative_ReturnsZero)
 TEST_F(ReadableFileTests, RandomRead_DownloadReturnsNegative_ReturnsZero)
 {
     // Arrange
-    constexpr uint64_t offset = 50;
-    constexpr size_t bytesToRead = 100;
+    constexpr int64_t offset = 50;
+    constexpr int64_t bytesToRead = 100;
     std::vector<char> buffer(bytesToRead);
 
     EXPECT_CALL(*m_blobClient, DownloadTo(::testing::A<std::span<char>>(), offset, bytesToRead))
-        .WillOnce(Return(-1)); // Simulate error
+      .WillOnce(Return(-1)); // Simulate error
 
     ReadableFileImpl file{ "test.sst", m_blobClient, nullptr };
 
-    // Act
-    const auto bytesRead = file.RandomRead(offset, bytesToRead, buffer.data());
+ // Act
+  const auto bytesRead = file.RandomRead(offset, bytesToRead, buffer.data());
 
-    // Assert
+  // Assert
     EXPECT_EQ(0, bytesRead);
 }
 
 TEST_F(ReadableFileTests, SequentialRead_EmptyBlob_ReturnsZero)
 {
     // Arrange
-    constexpr uint64_t blobSize = 0;
-    std::vector<char> buffer(100);
+ constexpr uint64_t blobSize = 0;
+  std::vector<char> buffer(100);
 
     EXPECT_CALL(*m_blobClient, GetSize())
-        .WillOnce(Return(blobSize));
+      .WillOnce(Return(blobSize));
 
     ReadableFileImpl file{ "test.sst", m_blobClient, nullptr };
 
@@ -366,7 +367,7 @@ TEST_F(ReadableFileTests, RandomRead_EmptyBlob_ReturnsZero)
     std::vector<char> buffer(100);
 
     EXPECT_CALL(*m_blobClient, GetSize())
-        .WillOnce(Return(blobSize));
+  .WillOnce(Return(blobSize));
 
     ReadableFileImpl file{ "test.sst", m_blobClient, nullptr };
 
@@ -379,22 +380,22 @@ TEST_F(ReadableFileTests, RandomRead_EmptyBlob_ReturnsZero)
 
 TEST_F(ReadableFileTests, SequentialRead_InterleavedWithSkip_MaintainsCorrectOffset)
 {
-    // Arrange
-    constexpr size_t readSize = 50;
-    constexpr uint64_t skipAmount = 25;
+  // Arrange
+    constexpr int64_t readSize = 50;
+    constexpr int64_t skipAmount = 25;
     std::vector<char> buffer(readSize);
 
     EXPECT_CALL(*m_blobClient, DownloadTo(::testing::A<std::span<char>>(), 0, readSize))
         .WillOnce(Return(static_cast<int64_t>(readSize)));
 
     EXPECT_CALL(*m_blobClient, DownloadTo(::testing::A<std::span<char>>(), readSize + skipAmount, readSize))
-        .WillOnce(Return(static_cast<int64_t>(readSize)));
+  .WillOnce(Return(static_cast<int64_t>(readSize)));
 
     ReadableFileImpl file{ "test.sst", m_blobClient, nullptr };
 
     // Act
     [[maybe_unused]] const auto bytesRead1 = file.SequentialRead(readSize, buffer.data());
-    file.Skip(skipAmount);
+  file.Skip(skipAmount);
     [[maybe_unused]] const auto bytesRead2 = file.SequentialRead(readSize, buffer.data());
 
     // Assert
